Added controller config
