+++ public/apps/media.html
@@ -4,17 +4,18 @@
   <meta charset="utf-8">
   <title>3dsNode | Media</title>
   <meta name="viewport" content="width=1000,height=1000,initial-scale=1.001,user-scalable=no">
-  <link rel="icon" href="./logo/icon.ico">
+  <link rel="icon" href="../logo/icon.ico">
 </head>
 <body>
   <div id="top">
-	<iframe id="topscreen" src="./logo/logo.gif" width="400" height="215" frameborder="0"></iframe>
+	<iframe id="topscreen" src="../logo/logo.gif" width="400" height="215" frameborder="0"></iframe>
   </div>
   <div id="bottom">
   </div>
   <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/1.7.2/jquery.min.js"></script>
   <script src="/socket.io/socket.io.js"></script>
-  <script type="text/javascript" src="./js/media.js"></script>
-  <script type="text/javascript" src="./js/bottom.js"></script>
+  <script type="text/javascript" src="../js/media.js"></script>
+  <script type="text/javascript" src="../js/bottom.js"></script>
+  <script type="text/javascript" src="../js/appcss.js"></script>
 </body>
 </html>
